ATM interface
I've written the code using Swings in Java. The Page1 depicts the login page
Page2 depicts the Page for creating an account
Page3 depicts the login details where minimum amount of deposit is 10000/-
If we link this interface with the database, we can get the user info, save the user details and maintain security.